Mono dialogue created + QuestsAndDialoguesSO
Ways to go backwards and forwards in DialogueUI text added (see DialogueManager); In case a line in the array has [P] then it is removed and is assumed to be the player who is speaking, that is why for now i have the name switched to player.

Later on might be planning to make a QuestProgress script that would track every progress out there and the only way to retrieve specific progresses might be through unique code. Retrieving it might be crucial for switching out different dialogues, albeit it being a quest dialogue array or just the one for normal chatting.