Dialogue switching and button switching
Dialogue switching from quest only dialogues to default ones; it can only happen on the first line, on the other lines then 1,2,3 inputs change to check if it hit the last line of the array and whether it's a quest dialogue or not.
Additionally added Button layout system so for the player's perspective those buttons would change depending on which line the player is and what type of dialogue.