QuestLog
QuestLogTemplate prefab created, QuestLogItems  instantiates the templates, the template has a QuestLogItemUI which holds the logic of sorting itself using the given data and quest status.

Added an event system when quest starts in QuestManager and when the status of the quest changes. Additionally have made some too but they have to be worked on.
Now there is a bool in the QuestsAndDialogueSO that figures out if its a main quest or a side quest so the Quest Log could show which one it is.