UI above NPCs, Pop up Sound, Freeze
Implented sound to the UI pop up, now the quest order (for main quests) is also important for UIs above NPC heads, QuestUIController handles the logic of every UI it being side quest or main quest. In the DialogueManager script I have a listener that listens to the QuestManager if the quest cannot start due to the player not starting the 1st quest or the previous quest before it, so now IsMainQuest variable in the SO determines if a quest has an order to it or not.
Additionally cleaned up some code for better readability.